backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-engineio==4.8.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0
cors==1.0.1
aiofiles==23.2.1
